requirements.txt
requests
beautifulsoup4
lxml
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
python-dotenv 